update the UI for the matching pairs
