--- devU-client/src/components/pages/Multiplechoice/matchingTable.scss
@@ -1,121 +1,0 @@
-@import 'variables';
-
-// Root container
-.matching-table-container {
-  background: var(--surface);
-  padding: 2rem;
-  border-radius: var(--border-radius);
-  box-shadow: var(--shadow-md);
-  border: 1px solid #e0e0e0;
-  margin-bottom: 2rem;
-}
-
-// Instructions box
-.instructions-box {
-  background-color: var(--background);
-  padding: 1rem 1.5rem;
-  border-left: 5px solid var(--primary);
-  border-radius: var(--border-radius);
-  margin-bottom: 1.5rem;
-  color: var(--text-secondary);
-  font-size: 0.95rem;
-}
-
-// Matching grid
-.matching-columns {
-  display: flex;
-  justify-content: space-between;
-  gap: 2rem;
-
-  @media (max-width: 768px) {
-    flex-direction: column;
-  }
-}
-
-.prompts-column,
-.options-column,
-.status-column {
-  flex: 1;
-}
-
-.column-header {
-  font-weight: 700;
-  margin-bottom: 0.75rem;
-  color: var(--text-primary);
-  border-bottom: 1px solid #ddd;
-  padding-bottom: 0.5rem;
-}
-
-.prompt-item,
-.option-item {
-  background-color: var(--list-item-background);
-  border-radius: var(--border-radius);
-  padding: 0.75rem 1rem;
-  margin-bottom: 0.5rem;
-  box-shadow: var(--shadow-sm);
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  transition: background 0.2s ease;
-
-  &:hover {
-    background-color: var(--list-item-background-hover);
-  }
-
-  &.active {
-    background-color: var(--primary-light);
-    color: white;
-    font-weight: 600;
-  }
-
-  &.matched {
-    border-left: 5px solid var(--primary-dark);
-  }
-}
-
-.item-number {
-  font-weight: bold;
-  color: var(--text-secondary);
-  margin-right: 1rem;
-  min-width: 1rem;
-}
-
-.remove-answer-btn {
-  background: transparent;
-  border: none;
-  font-size: 1.2rem;
-  cursor: pointer;
-  color: var(--error);
-
-  &:hover {
-    color: darkred;
-  }
-}
-
-.status-item {
-  padding: 0.5rem 0;
-  font-size: 0.9rem;
-
-  .status-icon.correct {
-    color: var(--success);
-    font-weight: 600;
-  }
-
-  .status-icon.incorrect {
-    color: var(--error);
-    font-weight: 600;
-  }
-
-  .correct-answer {
-    display: block;
-    font-size: 0.8rem;
-    color: var(--text-secondary);
-  }
-
-  .status-unanswered {
-    color: var(--text-secondary);
-    font-style: italic;
-  }
-}
-
